API処理リファクタ
