+++ backend/app/controllers/api/v1/work_items_controller.rb
@@ -35,6 +35,7 @@ module Api
             technology: ['React', 'TypeScript'],
           },
         ].to_json
+
         render json: work_items
       end
     end
